Make tests run over Tycho
References to meta-models doesn't work properly using platform URIs or with complex relative
paths. The issue seems to be that somehow Tycho run a minimal platform which doesn't do all the
required registrations (e.g., for the OCL environment and the inclusion of all test projects).
The workaround has been to create a symlink to reference the meta-models folder in the efinder.core.tests project.

Command: mvn verify